Gave up on the UDP checksum calculation for now and just output zeros.  GNU Radio is now picking up the packets!!!
There do seem to be some dropped counts in the data.  Happens every frame or so.  It's really easy to see when you view the data at 20 cols wide with xxd.  Just look at the last hex nibble on the right.  All of the last hex letters should be the same all the way down.  If they change, look in the row it changed and you should see a missed count.  It's possible these missed counts are coming from GNU Radio (it claims to be dropping packets)